+++ app/src/main/java/com/infineon/esim/lpa/euicc/identive/IdentiveConnectionBroadcastReceiver.java
@@ -42,6 +42,7 @@ public class IdentiveConnectionBroadcastReceiver extends BroadcastReceiver {
     private final OnDisconnectCallback onDisconnectCallback;
 
     private static boolean hasBeenFreshlyAttached = false;
+    private static String lastReaderName;
 
     public IdentiveConnectionBroadcastReceiver(Context context, OnDisconnectCallback onDisconnectCallback) {
         this.context = context;
@@ -55,10 +56,15 @@ public class IdentiveConnectionBroadcastReceiver extends BroadcastReceiver {
 
         switch (intent.getAction()) {
             case UsbManager.ACTION_USB_DEVICE_ATTACHED:
+                UsbDevice usbDevice = intent.getParcelableExtra(UsbManager.EXTRA_DEVICE);
+                lastReaderName = usbDevice.getProductName();
+
+                Log.info(TAG,"USB reader \"" + lastReaderName + "\" attached.");
                 hasBeenFreshlyAttached = true;
-//                /* Do not directly initialize because of user prompt. Only in onResume method in the
-//                   activity (e.g. ProfileListActivity).
-//                 */
+
+                /* Do not directly initialize because of user prompt. Only in onResume method in the
+                   activity (e.g. ProfileListActivity).
+                 */
                 break;
             case UsbManager.ACTION_USB_DEVICE_DETACHED:
                 onDisconnectCallback.onDisconnect();
@@ -75,10 +81,14 @@ public class IdentiveConnectionBroadcastReceiver extends BroadcastReceiver {
         context.registerReceiver(this, filter);
     }
 
-    public static Boolean hasBeenFreshlyAttached() {
+    public static Boolean hasBeenFreshlyAttached() throws Exception {
         if(hasBeenFreshlyAttached) {
             hasBeenFreshlyAttached = false;
-            return true;
+            if(isValidReaderName(lastReaderName)) {
+                return true;
+            } else {
+                throw new Exception("Reader \"" + lastReaderName + "\" not supported.");
+            }
         } else {
             return false;
         }
@@ -91,9 +101,7 @@ public class IdentiveConnectionBroadcastReceiver extends BroadcastReceiver {
         for (UsbDevice device : deviceList.values()) {
             Log.debug(TAG, "USB device attached: " + device.getProductName());
 
-            if (device.getProductName().equals(IdentiveEuiccInterface.READER_NAME)) {
-                return true;
-            }
+            return isValidReaderName(device.getProductName());
         }
 
         return false;
@@ -101,5 +109,15 @@ public class IdentiveConnectionBroadcastReceiver extends BroadcastReceiver {
 
     public interface OnDisconnectCallback {
         void onDisconnect();
+    }
+
+    private static boolean isValidReaderName(String readerName) {
+        for(String validReaderName : IdentiveEuiccInterface.READER_NAMES) {
+            if (readerName.equals(validReaderName)) {
+                return true;
+            }
+        }
+
+        return false;
     }
 }