+++ app/src/main/java/com/infineon/esim/lpa/euicc/se/SeEuiccInterface.java
@@ -25,6 +25,7 @@ package com.infineon.esim.lpa.euicc.se;
 
 import android.content.Context;
 import android.content.pm.PackageManager;
+import android.os.Build;
 
 import com.infineon.esim.lpa.Application;
 import com.infineon.esim.lpa.euicc.base.EuiccConnection;
@@ -63,7 +64,11 @@ final public class SeEuiccInterface implements EuiccInterface {
         boolean isAvailable = false;
 
         if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.R) {
+            // Starting with Android R the existence of OMAPI can be checked
             isAvailable = packageManager.hasSystemFeature(PackageManager.FEATURE_SE_OMAPI_UICC);
+        } else if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+            // Starting with Android P the OMAPI is part of Android but cannot be checked (yet).
+            isAvailable = true;
         }
 
         Log.debug(TAG, "Checking if SE eUICC interface is available: " + isAvailable);