+++ core/src/main/java/com/infineon/esim/lpa/core/dtos/apdu/Apdu.java
@@ -29,10 +29,27 @@ public class Apdu {
     private static final String TAG = Apdu.class.getName();
 
     // Common status words
-    private static final String SUCCESS_STATUS_WORD = "9000";
+    private static final String SW_SUCCESS_STATUS_WORD =            "9000";
+    private static final String SW_SUCCESS_PROACTIVE_CMD_WAITING =  "91";
 
     public static boolean isSuccessResponse(String response) {
-        return response != null && response.endsWith(SUCCESS_STATUS_WORD);
+        if((response == null) || (response.length() < 4)) {
+            return false;
+        }
+
+        String statusWord = response.substring(response.length()-4);
+
+        return statusWord.equals(SW_SUCCESS_STATUS_WORD)
+                || statusWord.startsWith(SW_SUCCESS_PROACTIVE_CMD_WAITING);
+
+    }
+
+    public static String getStatusWord(String response) {
+        if(response.length() >= 4) {
+            return response.substring(response.length() - 4);
+        } else {
+            return null;
+        }
     }
 
     public static boolean isLastResponseSuccess(List<String> responses) {