+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+## 4.0.1 (2022-08-10)
+
+Features:
+  - Added support for two new Identiv USB readers
+    - Identive CLOUD 4700 F Dual Interface Reader
+    - Identiv uTrust 4701 F Dual Interface Reader
+  - Added error message for unknown USB readers
+
 ## 4.0.0 (2022-06-24)
 
 Initial release for distribution.